Rename `EitherService` -> `Either`
